soc2-policy-generator/requirements.txt
Flask==3.0.3
python-docx==1.1.2
Jinja2==3.1.4
PyYAML==6.0.2